docs: upgrade vocs 1.4.1 → 2.0.4
Vocs v2 is a ground-up rewrite on Waku + RSC + Vite 8. Migration changes:

- Bump vocs to ^2.0.4 and add vite ^8 + waku ^1.0.0-beta peer deps.
- Move pages/ → src/pages/ and lib/ → src/lib/ (v2 default srcDir is "src").
- Rename vocs.config.tsx → vocs.config.ts (v2 only loads .ts/.js/.mjs/.mts).
- Update HomePage import in pages/index.mdx from 'vocs/components' to 'vocs'.
- Update Sidebar type import in sidebar.ts from 'vocs' to 'vocs/config'.
- defineConfig now imported from 'vocs/config'; the head() callback API is removed in v2.
- Drop the deprecated theme.accentColor wrapper for a flat accentColor.
- Drop aiCta (replaced by the per-page showAskAi frontmatter flag).
- Add renderStrategy: 'full-static' so all pages prerender to HTML.
- Mark svgsequence.tsx and swaggerviewer.tsx as 'use client' (required for RSC).
- Re-implement per-page canonical/JSON-LD/robots/title in src/pages/_mdx-wrapper.tsx using MdxPageContext.use() + useRouter() from waku. vocs's dedupeHead lets the wrapper's later tags win over the built-in <Head>'s defaults.
- Gitignore src/pages.gen.ts (Waku-generated route manifest) and docs/.vocs cache.
