+++ docs/package.json
@@ -18,6 +18,8 @@
     "swagger-ui-react": "^5.26.0",
     "tailwindcss": "^4.2.2",
     "typescript": "^5.8.3",
-    "vocs": "^1.0.12"
+    "vite": "^8.0.0",
+    "vocs": "^2.0.4",
+    "waku": "^1.0.0-beta.0"
   }
 }
