+++ docs/src/lib/swaggerviewer.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import SwaggerUI from 'swagger-ui-react';
 import 'swagger-ui-react/swagger-ui.css';
