+++ docs/src/pages/actions/signup.mdx
