+++ docs/src/pages/apis/cold_storage.mdx
