+++ docs/src/pages/apis/hot_storage.mdx
