+++ docs/src/pages/components/auth.mdx
