+++ docs/src/pages/components/cold_storage/otp.mdx
