+++ docs/src/pages/components/hot_storage.mdx
