+++ docs/src/pages/index.mdx
@@ -4,7 +4,7 @@ description: OpenSigner is an open-source, self-hostable wallet key management s
 ---
 
 
-import { HomePage } from 'vocs/components'
+import { HomePage } from 'vocs'
 
 <h1 className="sr-only">OpenSigner: open-source, non-custodial wallet key management</h1>
