+++ docs/src/pages/introduction/about.mdx
